openclawclone/requirements.txt
# Core Framework
langgraph
langchain-core
langchain-openai

# Telegram Bot
python-telegram-bot

# Utilities
python-dotenv
